+++ lib_/lib_to_make_libs.py
@@ -91,7 +91,7 @@ class ModuleFunction(BaseBuiltInFunction):
             return result
 
         # special built-in functions that needs the 'run' function (in nougaro.py) in their arguments
-        if method_dict["run_noug_dir_work_dir"]:
+        if method_dict["run_noug_dir"]:
             return_value = result.register(method(self, exec_context, run, noug_dir, work_dir))
 
             # if there is any error